+++ src/AngleSharp.Css/Dom/Internal/MediaFeature.cs
@@ -1,5 +1,6 @@
-﻿namespace AngleSharp.Css.Dom
+namespace AngleSharp.Css.Dom
 {
+    using AngleSharp.Css.Values;
     using AngleSharp.Text;
     using System;
     using System.IO;
@@ -13,8 +14,9 @@
 
         private readonly Boolean _min;
         private readonly Boolean _max;
-        private readonly String _name;
+        private readonly ICssValue _name;
         private readonly ICssValue _value;
+        private readonly String _op;
 
         #endregion
 
@@ -26,17 +28,28 @@
 
         internal MediaFeature(String name, ICssValue value)
         {
-            _name = name;
+            _name = new CssAnyValue(name);
             _value = value;
             _min = name.StartsWith("min-");
             _max = name.StartsWith("max-");
+        }
+
+        internal MediaFeature(String name, ICssValue value, String op)
+            : this(new CssAnyValue(name), value, op)
+        {}
+
+        internal MediaFeature(ICssValue name, ICssValue value, String op)
+        {
+            _name = name;
+            _value = value;
+            _op = op;
         }
 
         #endregion
 
         #region Properties
 
-        public String Name => _name;
+        public String Name => _name?.CssText ?? String.Empty;
 
         public Boolean IsMinimum => _min;
 
@@ -53,9 +66,16 @@
         public void ToCss(TextWriter writer, IStyleFormatter formatter)
         {
             writer.Write(Symbols.RoundBracketOpen);
-            writer.Write(_name);
+            writer.Write(Name);
 
-            if (_value != null)
+            if (_op is not null)
+            {
+                writer.Write(" ");
+                writer.Write(_op);
+                writer.Write(" ");
+                writer.Write(Value);
+            }
+            else if (_value is not null)
             {
                 writer.Write(": ");
                 writer.Write(Value);