+++ src/AngleSharp.Css/Parser/Micro/MediumParser.cs
@@ -1,6 +1,7 @@
 namespace AngleSharp.Css.Parser
 {
     using AngleSharp.Css.Dom;
+    using AngleSharp.Css.Values;
     using AngleSharp.Text;
     using System;
     using System.Collections.Generic;
@@ -23,92 +24,27 @@ namespace AngleSharp.Css.Parser
         /// </summary>
         public static CssMedium ParseMedium(this StringSource source, IFeatureValidatorFactory factory)
         {
-            source.SkipSpacesAndComments();
-            var ident = source.ParseMediumIdent();
-            var inverse = false;
-            var exclusive = false;
-            var type = String.Empty;
-
-            if (ident != null)
-            {
-                if (ident.Isi(CssKeywords.Not))
-                {
-                    inverse = true;
-                    source.SkipSpacesAndComments();
-                    ident = source.ParseMediumIdent();
-                }
-                else if (ident.Isi(CssKeywords.Only))
-                {
-                    exclusive = true;
-                    source.SkipSpacesAndComments();
-                    ident = source.ParseMediumIdent();
-                }
-            }
-
-            if (ident != null)
-            {
-                type = ident;
-                source.SkipSpacesAndComments();
-                var position = source.Index;
-                ident = source.ParseMediumIdent();
-
-                if (ident == null || !ident.Isi(CssKeywords.And))
-                {
-                    source.BackTo(position);
-                    return new CssMedium(type, inverse, exclusive);
-                }
-
-                source.SkipSpacesAndComments();
-            }
-
-            var features = new List<IMediaFeature>();
-
-            do
-            {
-                var start = source.Current;
-                source.SkipCurrentAndSpaces();
-                var feature = ParseFeature(source);
-                var end = source.Current;
-
-                if (feature == null || 
-                    start != Symbols.RoundBracketOpen || 
-                    end != Symbols.RoundBracketClose)
-                {
-                    return null;
-                }
-
-                var validator = factory?.Create(feature.Name);
-                feature.AssociateValidator(validator);
-                features.Add(feature);
-                source.SkipCurrentAndSpaces();
-                var position = source.Index;
-                ident = source.ParseMediumIdent();
-
-                if (ident == null || !ident.Isi(CssKeywords.And))
-                {
-                    source.BackTo(position);
-                    break;
-                }
-
-                source.SkipSpacesAndComments();
-            }
-            while (!source.IsDone);
-
-            return new CssMedium(type, inverse, exclusive, features);
-        }
-
-        /// <summary>
-        /// Parses a medium value.
-        /// </summary>
-        public static CssMedium ParseMediumNew(this StringSource source, IFeatureValidatorFactory factory)
-        {
+            // Syntax:
             // <media-query> = <media-condition> | <media-type>
+
             source.SkipSpacesAndComments();
-            return source.ParseMediaCondition() ?? source.ParseMediaType();
+            var medium = source.ParseMediaCondition() ?? source.ParseMediaType();
+
+            if (medium is not null)
+            {
+                foreach (var feature in medium.Features)
+                {
+                    var validator = factory?.Create(feature.Name);
+                    feature.AssociateValidator(validator);
+                }
+            }
+
+            return medium;
         }
 
         private static CssMedium ParseMediaCondition(this StringSource source)
         {
+            // Syntax:
             // <media-condition> = <media-not> | <media-in-parens> [ <media-and>* | <media-or>* ]
 
             var medium = source.ParseMediaNot();
@@ -117,10 +53,15 @@ namespace AngleSharp.Css.Parser
             {
                 medium = source.ParseMediaInParens();
 
-                if (medium != null)
+                if (medium is not null)
                 {
-                    var other = source.ParseMediaAnd();
-                    //TODO
+                    source.SkipSpacesAndComments();
+                    var other = source.ParseMediaConnectorMultiple(CssKeywords.And) ?? source.ParseMediaConnectorMultiple(CssKeywords.Or);
+
+                    if (other is not null)
+                    {
+                        medium = new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features.Concat(other.Features), other.Connector);
+                    }
                 }
             }
 
@@ -129,6 +70,7 @@ namespace AngleSharp.Css.Parser
 
         private static CssMedium ParseMediaConditionWithoutOr(this StringSource source)
         {
+            // Syntax:
             // <media-condition-without-or> = <media-not> | <media-in-parens> <media-and>*
 
             var medium = source.ParseMediaNot();
@@ -142,14 +84,14 @@ namespace AngleSharp.Css.Parser
                     do
                     {
                         source.SkipSpacesAndComments();
-                        var other = source.ParseMediaAnd();
+                        var other = source.ParseMediaConnector(CssKeywords.And);
 
                         if (other is null)
                         {
                             break;
                         }
 
-                        medium = new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features.Concat(other.Features));
+                        medium = new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features.Concat(other.Features), medium.Connector);
                     }
                     while (!source.IsDone);
                 }
@@ -160,38 +102,39 @@ namespace AngleSharp.Css.Parser
 
         private static CssMedium ParseMediaNot(this StringSource source)
         {
+            // Syntax:
             // <media-not> = not <media-in-parens>
 
-            if (source.IsIdentifier("not"))
+            var pos = source.Index;
+
+            if (source.IsIdentifier(CssKeywords.Not))
             {
-                var pos = source.Index;
-                source.ParseIdent();
-                source.SkipCurrentAndSpaces();
+                source.SkipSpacesAndComments();
                 var medium = source.ParseMediaInParens();
 
                 if (medium is not null)
                 {
                     source.SkipSpacesAndComments();
-                    return new CssMedium(medium.Type, !medium.IsInverse, medium.IsExclusive, medium.Features);
+                    return new CssMedium(medium.Type, !medium.IsInverse, medium.IsExclusive, medium.Features, medium.Connector);
                 }
-
-                source.BackTo(pos);
             }
 
+            source.BackTo(pos);
             return null;
         }
 
         private static CssMedium ParseMediaInParens(this StringSource source)
         {
+            // Syntax:
             // <media-in-parens> = ( <media-condition> ) | <media-feature> | <general-enclosed>
 
-            if (source.Current == '(')
+            if (source.Current == Symbols.RoundBracketOpen)
             {
                 var pos = source.Index;
                 source.SkipCurrentAndSpaces();
                 var medium = source.ParseMediaCondition();
 
-                if (medium is not null && source.Current == ')')
+                if (medium is not null && source.Current == Symbols.RoundBracketClose)
                 {
                     source.SkipCurrentAndSpaces();
                     return medium;
@@ -203,72 +146,107 @@ namespace AngleSharp.Css.Parser
             return source.ParseMediaFeature() ?? source.ParseGeneralEnclosed();
         }
 
-        private static CssMedium ParseMediaAnd(this StringSource source)
+        private static CssMedium ParseMediaConnectorMultiple(this StringSource source, String connector)
         {
-            // <media-and> = and <media-in-parens>
+            // Syntax:
+            // [ <media-and>* | <media-or>* ]
 
-            if (source.IsIdentifier("and"))
+            var part = source.ParseMediaConnector(connector);
+
+            if (part is not null)
             {
-                var pos = source.Index;
-                source.ParseIdent();
-                source.SkipCurrentAndSpaces();
-                var medium = source.ParseMediaInParens();
+                var features = Enumerable.Empty<IMediaFeature>();
 
-                if (medium is not null)
+                while (part is not null)
                 {
                     source.SkipSpacesAndComments();
-                    return new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features, "and");
+                    features = features.Concat(part.Features);
+                    part = source.ParseMediaConnector(connector);
                 }
 
-                source.BackTo(pos);
+                return new CssMedium(String.Empty, false, false, features, connector);
             }
 
             return null;
         }
 
-        private static CssMedium ParseMediaOr(this StringSource source)
+        private static CssMedium ParseMediaConnector(this StringSource source, String connector)
         {
+            // Syntax:
+            // <media-and> = and <media-in-parens>
+            // or
             // <media-or> = or <media-in-parens>
 
-            if (source.IsIdentifier("or"))
+            var pos = source.Index;
+
+            if (source.IsIdentifier(connector))
             {
-                var pos = source.Index;
-                source.ParseIdent();
-                source.SkipCurrentAndSpaces();
+                source.SkipSpacesAndComments();
                 var medium = source.ParseMediaInParens();
 
                 if (medium is not null)
                 {
                     source.SkipSpacesAndComments();
-                    return new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features, "or");
+                    return new CssMedium(medium.Type, medium.IsInverse, medium.IsExclusive, medium.Features, connector);
                 }
-
-                source.BackTo(pos);
             }
 
+            source.BackTo(pos);
             return null;
         }
 
         private static CssMedium ParseMediaType(this StringSource source)
         {
+            // Syntax:
             // <media-type> = [ not | only ]? <ident> [ and <media-condition-without-or> ]?
-            //TODO
+
+            var pos = source.Index;
+            var inverse = source.IsIdentifier(CssKeywords.Not);
+            source.SkipSpacesAndComments();
+            var only = !inverse && source.IsIdentifier(CssKeywords.Only);
+            source.SkipSpacesAndComments();
+            var type = source.ParseIdent();
+            source.SkipSpacesAndComments();
+
+            if (type is not null)
+            {
+                if (!source.IsIdentifier(CssKeywords.And))
+                {
+                    return new CssMedium(type, inverse, only);
+                }
+
+                source.SkipSpacesAndComments();
+                var features = source.ParseMediaConditionWithoutOr();
+
+                if (features is not null)
+                {
+                    return new CssMedium(type, inverse, only, features.Features, features.Connector);
+                }
+
+            }
+
+            source.BackTo(pos);
             return null;
         }
 
         private static CssMedium ParseMediaFeature(this StringSource source)
         {
+            // Syntax:
             // <media-feature> = ( [ <mf-plain> | <mf-boolean> | <mf-range> ] )
 
-            if (source.Current == '(')
+            if (source.Current == Symbols.RoundBracketOpen)
             {
+                var pos = source.Index;
+                source.SkipCurrentAndSpaces();
                 var feature = source.ParseMediaFeaturePlain() ?? source.ParseMediaFeatureBoolean() ?? source.ParseMediaFeatureRange();
 
-                if (feature is not null && source.Current == ')')
+                if (feature is not null && source.Current == Symbols.RoundBracketClose)
                 {
                     source.SkipCurrentAndSpaces();
-                    return new CssMedium("", false, false, Enumerable.Repeat(feature, 1), "and");
+                    return new CssMedium(String.Empty, false, false, Enumerable.Repeat(feature, 1), CssKeywords.And);
                 }
+
+                source.BackTo(pos);
             }
 
             return null;
@@ -276,6 +254,7 @@ namespace AngleSharp.Css.Parser
 
         private static CssMedium ParseGeneralEnclosed(this StringSource source)
         {
+            // Syntax:
             // <general-enclosed> = [ <function-token> <any-value> ) ] | ( <ident> <any-value> )
             //TODO
             return null;
@@ -283,12 +262,16 @@ namespace AngleSharp.Css.Parser
 
         private static MediaFeature ParseMediaFeaturePlain(this StringSource source)
         {
-            // <mf-plain> = <ident> : <mf-value>
+            // Syntax:
+            // mf-plain> = <ident> : <mf-value>
+
             var pos = source.Index;
             var ident = source.ParseIdent();
+            source.SkipSpacesAndComments();
 
-            if (ident is not null && source.Current == ':')
+            if (ident is not null && source.Current == Symbols.Colon)
             {
+                source.SkipCurrentAndSpaces();
                 var value = source.ParseMediaFeatureValue();
 
                 if (value is not null)
@@ -296,16 +279,16 @@ namespace AngleSharp.Css.Parser
                     source.SkipSpacesAndComments();
                     return new MediaFeature(ident, value);
                 }
-
-                source.BackTo(pos);
             }
 
+            source.BackTo(pos);
             return null;
         }
 
         private static MediaFeature ParseMediaFeatureBoolean(this StringSource source)
         {
             // <mf-boolean> = <ident>
+
             var ident = source.ParseIdent();
 
             if (ident is not null)
@@ -319,51 +302,95 @@ namespace AngleSharp.Css.Parser
 
         private static MediaFeature ParseMediaFeatureRange(this StringSource source)
         {
-            // <mf-range> = <ident> '<' '='? | '>' '='? | '=' <mf-value>
+            // <mf-range> =
+            //       <ident> '<' '='? | '>' '='? | '=' <mf-value>
             //     | <mf-value> '<' '='? | '>' '='? | '=' <ident>
             //     | <mf-value> '<' '='? <ident> '<' '='? <mf-value>
             //     | <mf-value> '>' '='? <ident> '>' '='? <mf-value>
-            //TODO
+
+            var pos = source.Index;
+            var name = source.ParseIdentAsValue() ?? source.ParseMediaFeatureValue();
+
+            if (name is not null)
+            {
+                source.SkipSpacesAndComments();
+                var op = "";
+
+                if (source.Current == Symbols.LessThan || source.Current == Symbols.GreaterThan)
+                {
+                    var c = source.Current;
+
+                    if (source.Next() == Symbols.Equality)
+                    {
+                        op = $"{c}=";
+                        source.SkipCurrentAndSpaces();
+                    }
+                    else
+                    {
+                        op = $"{c}";
+                        source.SkipSpacesAndComments();
+                    }
+                }
+                else if (source.Current == Symbols.Equality)
+                {
+                    op = "=";
+                    source.SkipCurrentAndSpaces();
+                }
+                else
+                {
+                    source.BackTo(pos);
+                    return null;
+                }
+
+                if (name is Identifier)
+                {
+                    var value = source.ParseMediaFeatureValue();
+
+                    if (value is not null)
+                    {
+                        return new MediaFeature(name, value, op);
+                    }
+                }
+                else
+                {
+                    var value = source.ParseIdentAsValue();
+
+                    if (value is not null)
+                    {
+                        return new MediaFeature(name, value, op);
+                    }
+                }
+
+                source.BackTo(pos);
+            }
+
             return null;
         }
 
         private static ICssValue ParseMediaFeatureValue(this StringSource source)
         {
+            // Syntax:
             // <mf-value> = <number> | <dimension> | <ident> | <ratio>
-            //TODO
-            var ident = source.ParseNumber() ?? source.ParseIdent() ?? source.ParseUnit() ?? source.ParseRatio();
-            return null;
-        }
 
-        private static String ParseMediumIdent(this StringSource source)
-        {
             var ident = source.ParseIdent();
-            var current = source.Current;
-
-            while (current == '&' || current == '#' || current == ';' || current == '-')
+            
+            if (ident is not null)
             {
-                ident = null;
-                current = source.Next();
+                return new Identifier(ident);
             }
 
-            return ident;
-        }
+            var ratio = source.ParseRatio();
 
-        private static IMediaFeature ParseFeature(StringSource source)
-        {
-            var name = source.ParseMediumIdent();
-            var value = default(String);
-
-            if (name != null)
+            if (ratio.HasValue)
             {
+                return ratio.Value;
+            }
 
-                if (source.Current == Symbols.Colon)
-                {
-                    source.SkipCurrentAndSpaces();
-                    value = source.TakeUntilClosed();
-                }
-
-                return new MediaFeature(name, value);
+            var unit = source.ParseAtomicExpression();
+            
+            if (unit is not null)
+            {
+                return unit;
             }
 
             return null;