+++ src/AngleSharp.Css/Parser/Micro/NumberParser.cs
@@ -1,5 +1,6 @@
 namespace AngleSharp.Css.Parser
 {
+    using AngleSharp.Css.Values;
     using AngleSharp.Text;
     using System;
     using System.Globalization;
@@ -30,7 +31,7 @@ namespace AngleSharp.Css.Parser
         /// <summary>
         /// Parses the ratio (double) value.
         /// </summary>
-        public static Double? ParseRatio(this StringSource source)
+        public static Ratio? ParseRatio(this StringSource source)
         {
             var pos = source.Index;
             var top = source.ParseNumber();
@@ -39,7 +40,7 @@ namespace AngleSharp.Css.Parser
 
             if (top.HasValue && bottom.HasValue && c == Symbols.Solidus)
             {
-                return top.Value / bottom.Value;
+                return new Ratio(top.Value, bottom.Value);
             }
 
             source.BackTo(pos);