+++ src/AngleSharp.Css/Values/Primitives/Label.cs
@@ -6,7 +6,7 @@ namespace AngleSharp.Css.Values
     /// <summary>
     /// Represents a CSS label ("string") value.
     /// </summary>
-    struct Label : ICssPrimitiveValue, IEquatable<Label>
+    public struct Label : ICssPrimitiveValue, IEquatable<Label>
     {
         #region Fields