+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released on tbd.
 
 - Updated to use AngleSharp 1.0
+- Updated media parsing to media L4 spec (#133)
 - Fixed issue when updating shorthands with invalid values (#129)
 - Fixed issue with appended EOF character in `CssText` (#123)
 - Fixed missing semicolon in `@page` rule (#135)